hf_space/requirements.txt
torch==2.2.0
torchvision==0.17.0
transformers>=4.40.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
gradio>=4.44.0
huggingface_hub>=0.23.0
